refactor(svelte): split interval-state test monolith into harness suites
Move createIntervalState composite tests out of a single ~1k-line file into
a shared harness plus three responsibility-based suites (construction/modes,
bounds editor, brush/config). No production API change — pure helpers already
live under interval/; keep the factory shell co-located.

Closes #432
